+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.0
+
+- Add retry function to prevent messages get lost
+- Rename Parameter
+
 ## 1.1.0
 
 - add /metrics, /healthz and /readiness